llm_gateway/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pydantic==2.11.7
pydantic-settings==2.10.1
httpx==0.28.1
google-genai==1.32.0
